+++ app/Http/Resources/CocktailBasicResource.php
@@ -24,6 +24,9 @@ class CocktailBasicResource extends JsonResource
             'slug' => $this->slug,
             'name' => $this->name,
             'short_ingredients' => $this->getIngredientNames(),
+            'image' => $this->when($this->relationLoaded('images'), function () {
+                return new ImageResource($this->getMainImage());
+            }),
         ];
     }
 }