+++ CHANGELOG.md
@@ -1,13 +1,29 @@
-# v5.0.3
-## Fixes
-- Optimized ancestors query execution
+# v5.1.0
+## New
+- Added `thumb_url` to `Image` schema
+- Added `image` to `CocktailBasic` and `IngredientBasic` schemas
+- Added POST `/bars/{id}/optimize` method
+    - Used to run some optional commands per bar like, recalculating ABV, refreshing bar search index, recalculating ingredient hierarchy
+- Added cocktail tag index
 
+## Fixes
+- Fixed `Menu` and `MenuRequest` schemas being out of sync
+- Fixed duplicated queries when fetching descendants
+
+# v5.0.4
+## Fixes
+- Fixed cast error when duplicating recipe
+
+# v5.0.3
 ## Fixes
 - Optimized ancestors query execution
 
 # v5.0.2
 ## Fixes
 - Added missing menu schema changes
+
+## Changes
+- Added route parameters type validation
 
 # v5.0.1
 ## Fixes